baseten_exp/tutorial_helper/requirements.txt
transformers>=4.44.0
torch>=2.1.0
torchvision>=0.16.0
opencv-python>=4.8.0
Pillow>=10.0.0
requests>=2.31.0
num2words>=0.5.12
decord>=0.6.0
accelerate>=0.20.0